examples/computer_use_agent/requirements.txt
arize-phoenix
arize-phoenix-evals
arize-phoenix-otel
beautifulsoup4==4.12.3
langchain==0.3.14
langchain-anthropic==0.3.0
langchain-community==0.3.14
langchain-core==0.3.30
langchain-text-splitters==0.3.5
langgraph==0.2.62
loguru==0.7.2
pandas==2.2.3
python-dotenv==1.0.1
tqdm==4.66.6
anthropic==0.44.0
gradio
openinference-instrumentation-anthropic
opentelemetry-sdk
opentelemetry-exporter-otlp